backend/requirements.txt
Django==5.2.5
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg[binary]==3.2.9
python-dotenv==1.0.1
requests==2.32.3

